requirements.txt
sentence-transformers>=2.2.0
scikit-learn>=1.2.0
numpy>=1.21.0
langdetect>=1.0.9
fastapi>=0.110.0
uvicorn>=0.23.0
matplotlib>=3.6.0
streamlit>=1.30.0
feedparser>=6.0.10
pyyaml>=6.0
